+++ trackgpt/__init__.py
@@ -1,6 +1,0 @@
-"""Initializes the track_gpt package."""
-
-from .analysts import LangChainAnalyst, LlamaIndexAnalyst
-from .version import __version__
-
-__all__ = ["__version__", "LangChainAnalyst", "LlamaIndexAnalyst"]